Prepare Multiple Main Demo
