+++ bp.sh
@@ -1,0 +1,4 @@
+if [ $1 = "contract.deploy" ]
+then
+  ./gradlew contractDeploy
+fi
