+++ src/main/java/com/klaytn/caver/boilerplate/contract/README.md
@@ -1,0 +1,6 @@
+# Boiler Plate for Contract
+> caver-java SDK를 사용해서 스마트 컨트랙트를 어떻게 배포하는지, 배포한 스마트 컨트랙트는 또 어떻게 사용하는지에 대해 파악할 수 있는 Boiler Plate입니다.
+> 
+
+## Deploy
+> Smart Contract를 caver-java를 통해 배포하는 과정을 다룹니다.
